// src/app/features/gestion/exports/enums/export-etat.enum.ts
export enum ExportEtatEnum {
  //TODO: verify IDs from backend when available
  EtatDesClientsAvecFacturesEtAvoirs = '1',
  EtatReleveDesEncours = '2',
  EtatDesPartenaires = '3',
  EtatCADesPrestations = '4',
  EtatListeDesControlesNonFactures = '5',
  EtatDesMandants = '6',
  EtatListeDesControles = '7',
  EtatJournalDesModificationsDesReglements = '8',
  EtatActiviteGlobaleDesClientsPro = '9',
  EtatJournalDesVentes = '10',
  EtatSyntheseFinanciereJournee = '11',
  EtatRecapitulatifDesCloturesEtEncaissements = '12',
  EtatProductionParAnneeImmatriculation = '14',
  EtatActiviteParControleurEtActiviteDuCentre = '15',
  EtatTauxDeContreVisite = '16',
}